+++ CHANGELOG.md
@@ -33,7 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
--   Fix Portuguese translation issues ([#74](https://github.com/brightlayer-ui/react-auth-shared/issues/74)).
+-   Fix Portuguese translation issues ([#74](https://github.com/etn-ccis/blui-react-auth-shared/issues/74)).
 
 ## v3.7.0 (April 26, 2022)
 
@@ -51,7 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
--   Fixed non-bold text on Forgot Password screen when using Portuguese translations ([#59](https://github.com/brightlayer-ui/react-auth-shared/issues/59)).
+-   Fixed non-bold text on Forgot Password screen when using Portuguese translations ([#59](https://github.com/etn-ccis/blui-react-auth-shared/issues/59)).
 
 ## v3.6.0 (December 9, 2021)